oshmem/memheap/base: Fix excessive memory consumption
Allocate only required number of segments.
At scale, each PE allocates 32 entries for every other PE which leads to
a high memory consumprion. This is the major contributor to the scaling
portion of the memory footprint.

This PR provides enables dynamic adjustment of the storage space based
on the actual number of segments.

Signed-off-by: Artem Y. Polyakov <artpol84@gmail.com>