oshmem: Reduce the service memory footprint
Current implementation is maintaining a static cache of 32 entries
per remote peer containing memory key information.
In practice, only 2-3 entries are used. At scale, however this leads
to a significant overhead.

Example:
* one cache entry is 40B
* with default 32 entries per peer it is 1280B per peer.
* on cluster with 1000 nodes and 40 PPN it results in 51MB per process
* Cumulatively, per node (given 40PPN) its 51MB x 40 = 2GB of overhead

This commit introduces dynamic cache that contains only required number
of entries. With 2 entries by default this leads to 16x reduction in the
cache memory footprint.

Future optimization of the cache entry layout is planned.
Will be PR'd as a follow-up.

Co-authored-by: Artem Y. Polyakov <artemp@nvidia.com>
Signed-off-by: Subhadeep Bhattacharya <subhadeepb@nvidia.com>
Signed-off-by: Artem Polyakov <artemp@nvidia.com>